+++ R/once.R
@@ -4,6 +4,7 @@
 #'
 #' @param . variable to set
 #' @param val the value to set for the variable
+#' @param envir environment where variables resides
 #'
 #' @return the variable set to the new variable, along with a class of once added to the output
 #'
@@ -66,14 +67,14 @@
 #'
 #' @export
 
-setOnce <- function(., val = 1L) {
-  if(!inherits(.,"once")){
+setOnce <- function(., val = 1L, envir = NULL) {
   .. <- substitute(.)
+  if(is.null(envir)) envir <- getEnvir(as.character(..))
+  if(!inherits(get(as.character(..),envir = envir),"once")){
   if (typeof(..) != "symbol") stop(paste0(.., " must be an object."))
   res <- val
   class(res) <- c('once',class(res))
-  assign(as.character(..), res, envir = getEnvir(as.character(..)))
+  assign(as.character(..), res, envir = envir)
   }
 }
-